Improvement to the order service file

+++ database/migrations/2023_10_30_162459_create_restaurant_branches_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
                 ->constrained()
                 ->cascadeOnDelete()
                 ->cascadeOnUpdate();
-                $table->float('deliveryprice');
+                $table->decimal('deliveryprice');
                 $table->integer('arrivaltime');
                 $table->string('image');
                 $table->boolean('status')->default(false);
